+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## x.x.x
 
+- Added whitelist and blacklist options to be able to select consumers instead
+  of having random consumers.
 - Added the port parameter to be able to be set a new port for the alternate
   upstream.